Remove -f
