+++ app/Main.hs
@@ -7,7 +7,6 @@ import SLED.Prelude
 import Blammo.Logging.Simple
 import SLED.App
 import SLED.Options
-import SLED.Report
 import SLED.Run
 
 main :: IO ()
@@ -16,8 +15,7 @@ main = do
   app <- App opts <$> newLoggerEnv
 
   flip runAppT app $ do
-    report <- getReportSuggestion oFormat
-    n <- runLsd oPath oResolver oChecks oFilter oExcludes report
+    n <- runLsd oPath oResolver oChecks oFilter oExcludes
     logDebug $ "Suggestions found" :# ["count" .= n]
 
     when (n /= 0 && not oNoExit) $ do
