+++ src/SLED/Run.hs
@@ -4,6 +4,8 @@ module SLED.Run
 
 import SLED.Prelude
 
+import Blammo.Logging.Colors
+import Blammo.Logging.Logger
 import SLED.Check
 import SLED.Checks
 import SLED.StackYaml
@@ -11,16 +13,21 @@ import SLED.StackageResolver
 import System.FilePath.Glob
 
 runLsd
-  :: (MonadUnliftIO m, MonadLogger m, MonadStackYaml m)
+  :: ( MonadUnliftIO m
+     , MonadLogger m
+     , MonadStackYaml m
+     , MonadReader env m
+     , HasLogger env
+     )
   => FilePath
   -> Maybe StackageResolver
   -> ChecksName
   -> Maybe Pattern
   -> [Pattern]
-  -> (Suggestion -> m ())
   -> m Int
-runLsd path mResolver checksName mFilter excludes report = do
+runLsd path mResolver checksName mFilter excludes = do
   StackYaml {..} <- loadStackYaml path
+  Colors {..} <- getColorsLogger
 
   let
     resolver = fromMaybe syResolver mResolver
@@ -35,10 +42,24 @@ runLsd path mResolver checksName mFilter excludes report = do
     details <- getExternalDetails resolver extraDep
 
     for (checksByName checksName) $ \check -> do
-      let mSuggestion = runCheck check details extraDep
-      mSuggestion <$ traverse_ report mSuggestion
+      with (runCheck check details extraDep) $ \Suggestion {..} ->
+        pushLoggerLn
+          $ case sAction of
+            Remove ->
+              green "Remove " <> " " <> magenta (extraDepToText sTarget)
+            ReplaceWith r ->
+              yellow "Replace"
+                <> " "
+                <> magenta (extraDepToText sTarget)
+                <> " with "
+                <> cyan (extraDepToText r)
+          <> "\n        ↳ "
+          <> sDescription
 
   pure $ length $ catMaybes $ concat results
 
 filterExcludes :: [Pattern] -> [ExtraDep] -> [ExtraDep]
 filterExcludes excludes = filter $ \e -> not $ any (`matchPattern` e) excludes
+
+with :: (Foldable t, Applicative f) => t a -> (a -> f b) -> f (t a)
+with x f = x <$ for_ x f
